requirements.txt
sanic
sanic_jinja2
werkzeug

asyncpg
aiosqlite
tortoise-orm
pydantic

asyncclick
